feat(components): creates basic address component
