+++ src/components/address/index.ts
@@ -1,0 +1,2 @@
+export * from './basic';
+export * from './copy';
